+++ softgym/envs/pour_water_amount.py
@@ -129,13 +129,13 @@ class PourWaterAmountPosControlEnv(PourWaterPosControlEnv):
         elif self.observation_mode in ['point_cloud', 'key_point']:
             if self.observation_mode == 'point_cloud':
                 particle_pos = np.array(pyflex.get_positions()).reshape([-1, 4])[:, :3].flatten()
-                pos = np.zeros(shape=self.particle_obs_dim, dtype=np.float)
+                pos = np.zeros(shape=self.particle_obs_dim, dtype=np.float32)
                 pos[:len(particle_pos)] = particle_pos
                 cup_state = np.array([self.glass_x, self.glass_y, self.glass_rotation, self.glass_dis_x, self.glass_dis_z, self.height,
                                   self.glass_distance + self.glass_x, self.poured_height, self.poured_glass_dis_x, self.poured_glass_dis_z,
                                   self.line_box_y, self.current_config['targe_amount']])
             else:
-                pos = np.empty(0, dtype=np.float)
+                pos = np.empty(0, dtype=np.float32)
 
                 cup_state = np.array([self.glass_x, self.glass_y, self.glass_rotation, self.glass_dis_x, self.glass_dis_z, self.height,
                                   self.glass_distance + self.glass_x, self.poured_height, self.poured_glass_dis_x, self.poured_glass_dis_z,
