+++ softgym/envs/rope_env.py
@@ -68,7 +68,7 @@ class RopeNewEnv(FlexEnv):
             return self.get_image(self.camera_height, self.camera_width)
         if self.observation_mode == 'point_cloud':
             particle_pos = np.array(pyflex.get_positions()).reshape([-1, 4])[:, :3].flatten()
-            pos = np.zeros(shape=self.particle_obs_dim, dtype=np.float)
+            pos = np.zeros(shape=self.particle_obs_dim, dtype=np.float32)
             pos[:len(particle_pos)] = particle_pos
         elif self.observation_mode == 'key_point':
             particle_pos = np.array(pyflex.get_positions()).reshape([-1, 4])[:, :3]
